# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Real state",
        "description": "API para gestionar propiedades inmobiliarias",
        "version": "1.0.0"
    },
    "paths": {
        "/api/personas": {
            "get": {
                "tags": [
                    "Persona"
                ],
                "summary": "Obtener todas las personas",
                "operationId": "8601618afc784c4736e52022f7f6f047",
                "responses": {
                    "200": {
                        "description": "Lista de personas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Persona"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Persona"
                ],
                "summary": "Crear una nueva persona",
                "operationId": "d016dca10e1a596471fe5aa5d4c99841",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Persona"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Persona creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Persona"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        },
        "/api/personas/{id}": {
            "get": {
                "tags": [
                    "Persona"
                ],
                "summary": "Obtener una persona por ID",
                "operationId": "9b70ed1d108dd682ec86b7c48f328884",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la persona",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Persona"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Persona no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Persona"
                ],
                "summary": "Actualizar una persona por ID",
                "operationId": "2a6e07dc32d26f218b13c0a2dc442fdf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la persona",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Persona"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Persona actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Persona"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Persona no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Persona"
                ],
                "summary": "Eliminar una persona por ID",
                "operationId": "8564ed4ed7e2d07b09bda4aac6184710",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la persona",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona eliminada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Persona eliminada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Persona no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Persona": {
                "title": "Persona",
                "description": "Modelo de Persona",
                "required": [
                    "nombre",
                    "email",
                    "direccion"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Juan Perez"
                    },
                    "email": {
                        "type": "string",
                        "example": "juan@example.com"
                    },
                    "telefono": {
                        "type": "string",
                        "example": 12345678
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiKeyAuth": {
                "type": "apiKey",
                "name": "X-RapidAPI-Key",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "apiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-RapidAPI-Key"
            }
        }
    ]
}